Changed my mind, no need to use lazy_static
