+++ Cargo.toml
@@ -21,6 +21,3 @@ optional = true
 [features]
 profiling_enabled = []
 generate_bindings=["bindgen"]
-
-[dependencies]
-lazy_static = "1.1.0"
